Move slug from middleware to helpers
